% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.est.R
\name{G.est}
\alias{G.est}
\title{compute the Gini ratio}
\usage{
G.est(type = "xtype")
}
\arguments{
\item{xtype}{the type of x, xtype=c("rlnorm","runif","rbinom")}
}
\description{
Let X be a non-negative random variable with u=E[X].For a random sample (x_1,x_2,...,x_n) from the distribution of X.

the Gini ratio is defined by
}
\examples{
\dontrun{
G.est("rlnorm")
}
}
