% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCbeta.R
\name{MCbeta}
\alias{MCbeta}
\title{use the Monte Carlo to estimate of the Beta(3, 3)cdf}
\usage{
MCbeta(x)
}
\arguments{
\item{x}{the number of points between 0 and 1}
}
\value{
return the a Monte Carlo estimate and pbeta function in R for each x.
}
\description{
Write a function to compute a Monte Carlo estimate of the Beta(3, 3) cdf

Compare the estimates with the values returned by the pbeta function in R
}
\examples{
\dontrun{
x_seq <- seq(.1,.9,by=.1)
F_seq <- numeric(length(x_seq))
P_seq <- numeric(length(x_seq))
for (i in 1:length(x_seq)) {
  F_seq[i] <- F(x_seq[i])  # using the F function to estimate the probability
  P_seq[i] <- pbeta(x_seq[i],3,3) # using the pbeta function in R to estimate the probability
}
print(round(rbind(x_seq, F_seq, P_seq), 3))
}
}
