% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{make a faster version of table()}
\usage{
table2(x, y)
}
\arguments{
\item{x}{is a vector}

\item{y}{is a vector}
}
\value{
a table
}
\description{
make a faster version of table() for the case of an input of two integer vectors with no missing values
}
\examples{
\dontrun{
a <- c(4, 5, 6)
table2(a, a)
}
}
