% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ray.R
\name{Ray}
\alias{Ray}
\title{generate samples from a Rayleigh distribution.}
\usage{
Ray(x, r = 10000)
}
\arguments{
\item{x}{is the number of samples}

\item{r}{is the number of random data by runif ()}
}
\value{
the vector of generating from the Rayleigh distribution.
}
\description{
Implement a function to generate samples from a Rayleigh distribution,using antithetic variables.
}
\examples{
\dontrun{
x <- seq(0.1, 2.5, length = 10)
samples <- Ray(x)
}
}
